Extract semantic reranking into new azure-cosmos-ai module
Create new standalone Maven module sdk/cosmos/azure-cosmos-ai with:
- CosmosAIClientBuilder, CosmosAIAsyncClient, CosmosAIClient public API
- InferenceService rewritten to use azure-core HttpPipeline
- SemanticRerankResult/SemanticRerankScore models with package-private setters
- InferenceResponseParser for JSON parsing (in models package)
- Unit tests (22 passing) and sample code

Remove all semantic rerank code from azure-cosmos:
- CosmosAsyncContainer.semanticRerank() and CosmosContainer.semanticRerank()
- CosmosAsyncClient.getOrCreateInferenceService() and close() cleanup
- ImplementationBridgeHelpers SemanticRerankResult/ScoreHelper classes
- ModelBridgeInternal.initialize() calls for SemanticRerank types
- Configs.java inference endpoint properties
- Old InferenceService, model classes, test, and sample files
