+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/CosmosContainer.java
@@ -4,7 +4,6 @@
 package com.azure.cosmos;
 
 import com.azure.cosmos.implementation.throughputControl.sdk.config.GlobalThroughputControlGroup;
-import com.azure.cosmos.implementation.inference.InferenceService;
 import com.azure.cosmos.implementation.RequestTimeoutException;
 import com.azure.cosmos.models.CosmosBatch;
 import com.azure.cosmos.models.CosmosBatchOperationResult;
@@ -28,7 +27,6 @@ import com.azure.cosmos.models.CosmosReadManyRequestOptions;
 import com.azure.cosmos.models.FeedRange;
 import com.azure.cosmos.models.FeedResponse;
 import com.azure.cosmos.models.PartitionKey;
-import com.azure.cosmos.models.SemanticRerankResult;
 import com.azure.cosmos.models.SqlQuerySpec;
 import com.azure.cosmos.models.ThroughputProperties;
 import com.azure.cosmos.models.ThroughputResponse;
@@ -307,43 +305,6 @@ public class CosmosContainer {
     private CosmosItemResponse<Object> blockDeleteItemResponse(Mono<CosmosItemResponse<Object>> deleteItemMono) {
         try {
             return deleteItemMono.block();
-        } catch (Exception ex) {
-            final Throwable throwable = Exceptions.unwrap(ex);
-            if (throwable instanceof CosmosException) {
-                throw (CosmosException) throwable;
-            } else {
-                throw ex;
-            }
-        }
-    }
-
-    private SemanticRerankResult blockSemanticRerankResponse(
-        Mono<SemanticRerankResult> semanticRerankResultMono,
-        Map<String, Object> options) {
-        // Derive the blocking timeout from the same option key used by InferenceService,
-        // so the sync API respects timeout_seconds just like the async path.
-        // Default is InferenceService.DEFAULT_REQUEST_TIMEOUT (120 s) if not supplied.
-        Duration blockTimeout = InferenceService.DEFAULT_REQUEST_TIMEOUT;
-        if (options != null) {
-            Object timeoutVal = options.get(InferenceService.OPTION_TIMEOUT_SECONDS);
-            if (timeoutVal instanceof Number) {
-                double seconds = ((Number) timeoutVal).doubleValue();
-                if (seconds > 0) {
-                    blockTimeout = Duration.ofMillis((long) (seconds * 1000));
-                }
-            }
-        }
-        try {
-            // Apply the same timeout semantics on the sync wrapper that InferenceService applies per
-            // HTTP attempt: convert a timeout to a typed RequestTimeoutException (408) instead of
-            // letting Reactor's raw RuntimeException("Timeout on blocking read") leak to the caller.
-            final Duration timeout = blockTimeout;
-            return semanticRerankResultMono
-                .timeout(timeout)
-                .onErrorMap(java.util.concurrent.TimeoutException.class,
-                    t -> new RequestTimeoutException(
-                        "Semantic rerank request timed out after " + timeout, (java.net.URI) null))
-                .block(timeout);
         } catch (Exception ex) {
             final Throwable throwable = Exceptions.unwrap(ex);
             if (throwable instanceof CosmosException) {
@@ -1077,40 +1038,6 @@ public class CosmosContainer {
         CosmosBulkExecutionOptions bulkOptions) {
 
         return this.blockBulkResponse(asyncContainer.executeBulkOperations(Flux.fromIterable(operations), bulkOptions));
-    }
-
-    /**
-     * Performs semantic reranking of documents using the inference service.
-     *
-     * <p><strong>Timeout:</strong> This method blocks with a default timeout of 120 seconds.
-     * To override, pass {@code "timeout_seconds"} (as a {@link Number}) in the {@code options} map,
-     * for example: {@code options.put("timeout_seconds", 30)}.
-     *
-     * <p><strong>Sync vs async timeout semantics:</strong> on the async API
-     * ({@code CosmosAsyncContainer.semanticRerank}) {@code timeout_seconds} is applied per HTTP attempt,
-     * so the total wall-clock can stretch across retries and backoff. On this sync API the same value
-     * is also applied as the {@link Mono#timeout(java.time.Duration) whole-operation deadline}
-     * around the inner {@link Mono}, so the call returns once the overall budget elapses regardless of
-     * how many retries the async path would otherwise perform. When the budget elapses the sync API
-     * surfaces a {@link CosmosException} with status {@code 408 Request Timeout} rather than a raw
-     * Reactor {@link RuntimeException}.
-     *
-     * @param rerankContext The query or context string used to score documents.
-     * @param documents The list of document strings to rerank.
-     * @param options Optional reranking parameters as a map. Supported keys include:
-     *                {@code return_documents} (Boolean) - whether to return document text in the response,
-     *                {@code top_k} (Integer) - maximum number of documents to return,
-     *                {@code batch_size} (Integer) - number of documents per batch,
-     *                {@code sort} (Boolean) - whether to sort results by relevance score,
-     *                {@code timeout_seconds} (Number) - per-request timeout in seconds (default: 120).
-     * @return The semantic rerank result.
-     */
-    @Beta(value = Beta.SinceVersion.V4_81_0, warningText = Beta.PREVIEW_SUBJECT_TO_CHANGE_WARNING)
-    public SemanticRerankResult semanticRerank(
-        String rerankContext,
-        List<String> documents,
-        Map<String, Object> options) {
-        return blockSemanticRerankResponse(this.asyncContainer.semanticRerank(rerankContext, documents, options), options);
     }
 
     /**
