+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/models/ModelBridgeInternal.java
@@ -769,7 +769,5 @@ public final class ModelBridgeInternal {
         PriorityLevel.initialize();
         SqlQuerySpec.initialize();
         SqlParameter.initialize();
-        SemanticRerankResult.initialize();
-        SemanticRerankScore.initialize();
     }
 }
