+++ sdk/cosmos/pom.xml
@@ -11,6 +11,7 @@
 
   <modules>
     <module>azure-cosmos</module>
+    <module>azure-cosmos-ai</module>
     <module>azure-cosmos-benchmark</module>
     <module>azure-cosmos-encryption</module>
     <module>azure-cosmos-spark_3</module>
